Convert all TypeScript files to JavaScript
TypeScript is more annoying than helpful. It takes too long to set up
TSLint and add types to everything, and the resulting code isn't more
succinct. So why bother? And Flow isn't any better — it's just the same,
only it takes longer to get ESLint working with Flow, and it's arguably
not as strict as TypeScript. I honestly think either Elm or BuckleScript
would be better.