Change node parents into relationships
There can be many different connections between different nodes, and
based on types of connections there may be different ways to draw them.
"Parents" lumps everything into one category and doesn't allow us to
switch based on the type of the connection. Hence, for — for instance —
acquisition events, we have to render a "hidden" event just so we can
draw a line to a place where there is no event. The system doesn't
really know they're acquisition events, but if it did, there would be no
need for such an event.