Add more events to AT&T timeline
Everything looks mostly good right now, but we currently have an issue
where the events in which AT&T creates BCA and then RCA acquires BCA are
in the wrong order (they should be flipped). The reordering logic should
be doing this, except that we have some conflicting rules and so the
logic is just ignoring those rules. In this case it is possible to
invert the last rule, but the system doesn't know it can do that.