Rewrite logic to reorder node groups
We reorder node groups (which you can think of as "tracks" that
represent the companies involved in the graph) by using the
relationships between nodes to come up with a set of rules for how one
or more sets of node groups should be ordered, and then successively
applying these rules to determine a final ordering.

For instance, if we have a set of rules like:

    1 < 2 < 3,
    1 < 5 < 3,
    0 < 4.

we would want to combine this set into a master set, and then we would
know how to arrange the node groups.

However, combining rules like this is ambiguous. For instance, this
combined rule would satisfy the first two rules above:

    1 < 2 < 5 < 3.

but then so would this one:

    1 < 5 < 2 < 3.

And where does the third rule go, at the beginning or at the end of this
combined rule?

The way the rule-applying algorithm currently works is by taking a
tuple (which represents a rule) and a working series of indices (the
combined rule), treating them as sequences and aligning them, and then
merging the two sequences together, prioritizing the new sequence over
the existing sequences for columns that have different indices in them.
For instance, say we represented the first rule above as:

    [1, 2, 3]

and then we wanted to merge this tuple into it:

    [1, 5, 3]

We would first align these sequences like so:

    [1, 2, 3]
    [1, 5, 3]

then apply the bottom sequence over the top sequence, so that we end up
with:

    [1, 5, 2, 3]

If the sequences were different lengths, we'd add null values to align
them before merging them. So these two sequences:

    [1, 2, 3]
    [4, 3, 7]

would become:

    [   1, 2, 3, null]
    [null, 4, 3, 7]

and then the merged sequence would be:

    [1, 4, 2, 3, 7]

However, this algorithm breaks down when asked to apply a rule that
contradicts a previous rule. For instance, let's say we have this set of
rules:

    1 < 2 < 3,
    4 < 5,
    5 < 2.

If we were to use our algorithm, we would apply the second rule to the
first to get:

    [1, 2, 3, 4, 5]

but then we'd try to align this sequence with this one:

    [5, 2]

That simply can't be done.

This commit fixes this problem by representing the combined rule not as
a flat list but as a nested set structure, with arrays inside arrays
(what we call a RuleTree). Let's look at the first example again with
this new way. We actually start off with an empty RuleTree:

    RuleTree()

As we apply rules (what we now call RuleChains) we inject them into the
RuleTree in places where they intersect. So at first we'd apply
[1, 2, 3] and get:

    RuleTree(RuleChain(1, 2, 3))

Now we apply [1, 5, 3] and realize that the last place it intersects the
tree is at the 3 (the 1 is already there). So we replace the 3 with
[5, 3] and we get:

    RuleTree(RuleChain(1, 2, RuleChain(5, 3)))

Finally we apply [0, 4] and get:

    RuleTree(
      RuleChain(1, 2, RuleChain(5, 3)),
      RuleChain(0, 4)
    )

Then we flatten it to get:

    [1, 2, 5, 3, 0, 4]

In the second example, we start by applying [1, 2, 3]:

    RuleTree(RuleChain(1, 2, 3))

Then [4, 3, 7]:

    RuleTree(RuleChain(1, 2, RuleChain(4, 3, 7)))

Then flatten to get:

    [1, 2, 4, 3, 7]

In the third example, we start by again applying [1, 2, 3]:

    RuleTree(RuleChain(1, 2, 3))

Then [4, 5]:

    RuleTree(
      RuleChain(1, 2, 3),
      RuleChain(4, 5)
    )

Finally we go to apply [5, 2]. At first it may seem like nothing can be
done, until we realize we can simply flip the rule chains:

    RuleTree(
      RuleChain(4, 5),
      RuleChain(1, 2, 3)
    )

And now we flatten to get:

    [4, 5, 1, 2, 3]

The astute reader will observe that this still satisifies the following
rules:

    1 < 2 < 3,
    4 < 5,
    5 < 2.